Refactor cloud function into multiple files part 2 - networks.py
