+++ examples/cloud-operations/network-dashboard/cloud-function/metrics/instances.py
@@ -3,7 +3,8 @@ from collections import defaultdict
 from google.protobuf import field_mask_pb2
 from google.cloud import asset_v1
 #TODO: Workaround
-import main as networks
+import main as main
+from . import metrics, networks
 
 def get_gce_instance_dict(config: dict):
   '''
@@ -35,7 +36,7 @@ def get_gce_instance_dict(config: dict):
   return gce_instance_dict
 
 
-def get_gce_instances_data(metrics_dict, gce_instance_dict, limit_dict, monitored_projects: list[str]):
+def get_gce_instances_data(config, metrics_dict, gce_instance_dict, limit_dict):
   '''
     Gets the data for GCE instances per VPC Network and writes it to the metric defined in instance_metric.
 
@@ -46,13 +47,15 @@ def get_gce_instances_data(metrics_dict, gce_instance_dict, limit_dict, monitore
       Returns:
         gce_instance_dict
   '''
+  # TODO: move out
   # Existing GCP Monitoring metrics for GCE instances
   metric_instances_limit = "compute.googleapis.com/quota/instances_per_vpc_network/limit"
 
-  for project in monitored_projects:
-    network_dict = networks.get_networks(project)
+  for project in config["monitored_projects"]:
+    network_dict = networks.get_networks(config, project)
 
-    current_quota_limit = networks.get_quota_current_limit(f"projects/{project}",
+    #TODO
+    current_quota_limit = main.get_quota_current_limit(f"projects/{project}",
                                                   metric_instances_limit)
     if current_quota_limit is None:
       print(
@@ -60,27 +63,24 @@ def get_gce_instances_data(metrics_dict, gce_instance_dict, limit_dict, monitore
       )
 
     # TODO: workaround
-    current_quota_limit_view = networks.customize_quota_view(current_quota_limit)
+    current_quota_limit_view = main.customize_quota_view(current_quota_limit)
 
     for net in network_dict:
       # TODO: workaround
-      networks.set_limits(net, current_quota_limit_view, limit_dict)
+      main.set_limits(net, current_quota_limit_view, limit_dict)
 
       usage = 0
       if net['self_link'] in gce_instance_dict:
         usage = gce_instance_dict[net['self_link']]
 
-       # TODO: workaround
-      networks.write_data_to_metric(
-          project, usage, metrics_dict["metrics_per_network"]
+      metrics.write_data_to_metric(
+          config,project, usage, metrics_dict["metrics_per_network"]
           ["instance_per_network"]["usage"]["name"], net['network_name'])
-       # TODO: workaround
-      networks.write_data_to_metric(
-          project, net['limit'], metrics_dict["metrics_per_network"]
+      metrics.write_data_to_metric(
+          config, project, net['limit'], metrics_dict["metrics_per_network"]
           ["instance_per_network"]["limit"]["name"], net['network_name'])
-       # TODO: workaround
-      networks.write_data_to_metric(
-          project, usage / net['limit'], metrics_dict["metrics_per_network"]
+      metrics.write_data_to_metric(
+          config, project, usage / net['limit'], metrics_dict["metrics_per_network"]
           ["instance_per_network"]["utilization"]["name"], net['network_name'])
 
     print(f"Wrote number of instances to metric for projects/{project}")
